Moved gitstat and tarutil to internal
