+++ internal/tarutil/dir_test.go
@@ -4,7 +4,7 @@ import (
 	"archive/tar"
 	"bytes"
 	"io"
-	"os"
+	"io/fs"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -29,7 +29,9 @@ func TestDir(t *testing.T) {
 
 type antiBarIgnorer struct{}
 
-func (i antiBarIgnorer) Ignore(info os.FileInfo) bool { return info.Name() == "bar" }
+func (i antiBarIgnorer) Ignore(path string, info fs.FileInfo) bool {
+	return info.Name() == "bar"
+}
 
 func TestDirIgnore(t *testing.T) {
 	var buf bytes.Buffer
